requirements.txt
Django==3.2
djangorestframework==3.13.1
django-cors-headers==3.11.0
django-3-jet==1.0.8
django-ckeditor==6.4.1
Pillow==9.1.1
django-cleanup==6.0.0
mixer==7.2.2
django-solo==2.0.0
drf-yasg==1.20.0
psycopg2-binary==2.9.3
djangorestframework-simplejwt
python-decouple==3.8




